extract mpi feature and update some libraries
