+++ src/simulation/engines/teleportation_engine.rs
@@ -1,6 +1,6 @@
 use crate::simulation::id::Id;
-use crate::simulation::messaging::communication::communicators::SimCommunicator;
 use crate::simulation::messaging::communication::message_broker::NetMessageBroker;
+use crate::simulation::messaging::communication::SimCommunicator;
 use crate::simulation::messaging::events::EventsPublisher;
 use crate::simulation::simulation::Simulation;
 use crate::simulation::time_queue::TimeQueue;
