+++ src/simulation/simulation.rs
@@ -6,8 +6,8 @@ use std::rc::Rc;
 use crate::simulation::config::Config;
 use crate::simulation::engines::activity_engine::ActivityEngine;
 use crate::simulation::engines::leg_engine::LegEngine;
-use crate::simulation::messaging::communication::communicators::SimCommunicator;
 use crate::simulation::messaging::communication::message_broker::NetMessageBroker;
+use crate::simulation::messaging::communication::SimCommunicator;
 use crate::simulation::messaging::events::EventsPublisher;
 use crate::simulation::population::agent_source::{AgentSource, DefaultAgentSource};
 use crate::simulation::scenario::Scenario;
