+++ tests/test_adhoc_routing.rs
@@ -3,7 +3,7 @@ use std::path::PathBuf;
 use rust_q_sim::simulation::config::CommandLineArgs;
 use rust_q_sim::simulation::id::store_to_file;
 use rust_q_sim::simulation::logging::init_std_out_logging;
-use rust_q_sim::simulation::messaging::communication::communicators::DummySimCommunicator;
+use rust_q_sim::simulation::messaging::communication::local_communicator::DummySimCommunicator;
 use rust_q_sim::simulation::network::global_network::Network;
 use rust_q_sim::simulation::population::population_data::Population;
 use rust_q_sim::simulation::vehicles::garage::Garage;
