+++ tests/test_simulation.rs
@@ -17,10 +17,9 @@ use rust_q_sim::simulation::controller::{
 };
 use rust_q_sim::simulation::id;
 use rust_q_sim::simulation::io::xml_events::XmlEventsWriter;
-use rust_q_sim::simulation::messaging::communication::communicators::{
-    ChannelSimCommunicator, SimCommunicator,
-};
+use rust_q_sim::simulation::messaging::communication::local_communicator::ChannelSimCommunicator;
 use rust_q_sim::simulation::messaging::communication::message_broker::NetMessageBroker;
+use rust_q_sim::simulation::messaging::communication::SimCommunicator;
 use rust_q_sim::simulation::messaging::events::{EventsPublisher, EventsSubscriber};
 use rust_q_sim::simulation::network::global_network::Network;
 use rust_q_sim::simulation::network::sim_network::SimNetworkPartition;
@@ -109,9 +108,7 @@ pub fn execute_sim<C: SimCommunicator + 'static>(
     info!(
         "Partitioning: Rank {rank}; Links {:?}; Nodes {:?}",
         &sim_net.get_link_ids(),
-        &sim_net
-            .nodes.keys().copied()
-            .collect::<HashSet<u64>>()
+        &sim_net.nodes.keys().copied().collect::<HashSet<u64>>()
     );
 
     let events = Rc::new(RefCell::new(EventsPublisher::new()));
